feat: Table variants static and infinite created
